v1.2.0: Forecast screen, pitch black mode, onboarding location lock & UI polish
- Add forecast details screen with weather condition tiles
- Add pitch black mode for AMOLED displays
- Auto-detect system accent color and device theme
- Move weather background to shared layer behind all pages
- Lock onboarding completion behind location permission grant
- Redesign onboarding to match app visual language
- Add gradient height slider and polish settings UI
- Refactor SettingsService to singleton pattern
- Bump version to 1.2.0+1

+++ android/app/src/main/kotlin/de/osemine/simple_weather/SimpleWeatherWidget.kt
@@ -20,9 +20,9 @@ class SimpleWeatherWidget : HomeWidgetProvider() {
                 setTextViewText(R.id.widget_temperature, temperature)
                 setTextViewText(R.id.widget_description, description)
 
-                val bgColor = widgetData.getInt("bgColor", -1)
-                val textColor = widgetData.getInt("textColor", -1)
-                val subtextColor = widgetData.getInt("subtextColor", -1)
+                val bgColor = widgetData.getLong("bgColor", -1L).toInt()
+                val textColor = widgetData.getLong("textColor", -1L).toInt()
+                val subtextColor = widgetData.getLong("subtextColor", -1L).toInt()
 
                 if (bgColor != -1) {
                     setInt(R.id.widget_root, "setBackgroundColor", bgColor)
